requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
requests==2.31.0
rundmc-whisper-nim==0.1.4
pyannote.audio==2.1.1
torch==2.2.2
numpy==1.26.4
pydantic==2.7.1